refactor(evolution): fold promotion gates into lifecycle
