+++ forgeflow_runtime/evolution/__init__.py
@@ -25,10 +25,16 @@ from forgeflow_runtime.evolution.doctor import (
     doctor_evolution_state as _doctor_evolution_state,
 )
 from forgeflow_runtime.evolution.lifecycle import (
+    PROMOTION_DECISION_DIR,
     adopt_example_rule,
     load_retired_rule_by_id as _load_retired_rule_by_id,
     load_retired_rules as _load_retired_rules,
     move_with_audit_rollback as _move_with_audit_rollback,
+    promotion_decision,
+    promotion_decision_path as _promotion_decision_path,
+    promotion_gate,
+    promotion_ready,
+    read_promotion_decision_records as _read_promotion_decision_records,
     retire_rule as _retire_rule,
     restore_rule as _restore_rule,
     rule_filename as _rule_filename,
@@ -40,14 +46,6 @@ from forgeflow_runtime.evolution.proposals import (
     proposal_approve,
     proposal_review,
     read_proposal_approval_records as _read_proposal_approval_records,
-)
-from forgeflow_runtime.evolution.promotion_gates import (
-    PROMOTION_DECISION_DIR,
-    promotion_decision,
-    promotion_decision_path as _promotion_decision_path,
-    promotion_gate,
-    promotion_ready,
-    read_promotion_decision_records as _read_promotion_decision_records,
 )
 from forgeflow_runtime.evolution.promotions import (
     PROMOTED_RULE_DIR,
