+++ tests/test_runtime_package_boundaries.py
@@ -11,7 +11,6 @@ EVOLUTION_MODULES = [
     "doctor",
     "promotions",
     "execution",
-    "promotion_gates",
     "proposals",
     "lifecycle",
 ]
